refactor: setting up general purpose test app
