+++ package.json
@@ -46,7 +46,6 @@
     "clean-webpack-plugin": "~0.1.19",
     "copy-webpack-plugin": "~4.5.1",
     "css-loader": "~0.28.11",
-    "directory-named-webpack-plugin": "~4.0.0",
     "duplicate-package-checker-webpack-plugin": "~3.0.0",
     "eslint-import-resolver-webpack": "~0.10.0",
     "eslint-loader": "~2.0.0",
