+++ test-app/src/components/HomeScreen/HomeScreen.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+import Icon from 'components/universal/Icon/Icon'
+
+const HomeScreen = () => (
+  <div>
+    <h1>Home screen!!</h1>
+    <p>
+      <Icon id="heart" className="text-danger" />
+    </p>
+  </div>
+)
+
+export default HomeScreen
