+++ test-app/src/index.html
@@ -6,7 +6,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
-  <title>InspireScript</title>
+  <title>Crystal ball 🔮</title>
 
   <!-- The very amazing polyfill.io -->
   <script src="https://ft-polyfill-service.herokuapp.com/v2/polyfill.min.js"></script>
