+++ test-app/src/index.js
@@ -9,7 +9,7 @@ import svgSymbolSpriteLoader from 'svg-symbol-sprite-loader'
 // ⚠️ Import application styles before application components so that base CSS
 // styles are included before component styles.
 import './styles/index.scss'
-import App from './components/App'
+import App from './components/App/App'
 import store from './dux/store'
 
 // Injects SVG symbol sprite into document from local storage if it exists,
