+++ test-app/src/styles/index.scss
@@ -1,6 +1,6 @@
 /*
- * InspireScript Project Styles Entry
- * The application styles for InspireScript should be driven as much as possible by
+ * Crystal Ball Project Styles Entry
+ * The application styles for Crystal ball should be driven as much as possible by
  * a common set of theme variables and classes. Components consumed by the project
  * that follow this pattern will match the appearance of the application as a whole.
  */
