+++ test/acceptance.spec.js
@@ -22,21 +22,12 @@ beforeAll(async () => {
 describe('application renders', () => {
   test('title renders', async () => {
     const title = await page.$eval('h1', h1 => h1.innerHTML)
-    expect(title).toEqual('InspireScript')
+    expect(title).toEqual('Crystal ball 🔮')
 
     const imgSrc = await page.$eval('img', image => image.src)
     expect(imgSrc).toEqual(
       expect.stringMatching(/\/static\/media\/Hood-dan-hawk\.[a-z0-9]+?\.jpg/)
     )
-  })
-
-  test('navigates to /magic-markdown', async () => {
-    const navigationPromise = page.waitForNavigation()
-    await page.click('a[href="/magic-markdown"]')
-    await navigationPromise
-
-    const title = await page.$eval('h1', h1 => h1.innerHTML)
-    expect(title).toEqual('🔮 InspireScript Magic Markdown')
   })
 })
